test-requirements.txt
hacking>=0.5.6,<0.8

coverage>=3.6
discover
fixtures>=0.3.14
python-subunit
sphinx>=1.1.2
oslo.sphinx
testrepository>=0.0.17
testscenarios>=0.4,<0.5
testtools>=0.9.32

posix_ipc>=0.9.8
MySQL-python
psycopg2
